Remove macros
