+++ Cargo.toml
@@ -28,12 +28,13 @@ serde_json = "1.0"
 sea-query-derive = { version = "0.1.0", path = "sea-query-derive", default-features = false, optional = true }
 bytes = { version = "1", optional = true }
 postgres-types = { version = "0", optional = true }
+sqlx = { version = "0", default-features = false, optional = true }
 
 [dev-dependencies]
 async-std = "1.8"
 
 [dev-dependencies.sqlx]
-version = "0.5"
+version = "0"
 default-features = false
 features = [
     "runtime-async-std-native-tls",
@@ -51,8 +52,8 @@ features = [
 default = [ "derive" ]
 derive = [ "sea-query-derive" ]
 postgres = [ "bytes", "postgres-types" ]
-sqlx-mysql = [ ]
-sqlx-postgres = [ ]
+sqlx-mysql = [ "sqlx/mysql" ]
+sqlx-postgres = [ "sqlx/postgres" ]
 
 [[example]]
 name = "derive"
