+++ src/driver/sqlx_mysql.rs
@@ -1,4 +1,6 @@
-#[macro_export]
+use sqlx::{query::Query, query::QueryAs, MySql, mysql::MySqlArguments};
+use crate::Value;
+
 macro_rules! bind_params_sqlx_mysql {
     ( $query:expr, $params:expr ) => {
         {
@@ -24,4 +26,15 @@ macro_rules! bind_params_sqlx_mysql {
             query
         }
     };
+}
+
+type SqlxQuery<'a> = Query<'a, MySql, MySqlArguments>;
+type SqlxQueryAs<'a, T> = QueryAs<'a, MySql, T, MySqlArguments>;
+
+pub fn bind_query<'a>(query: SqlxQuery<'a>, params: &'a [Value]) -> SqlxQuery<'a> {
+    bind_params_sqlx_mysql!(query, params)
+}
+
+pub fn bind_query_as<'a, T>(query: SqlxQueryAs<'a, T>, params: &'a [Value]) -> SqlxQueryAs<'a, T> {
+    bind_params_sqlx_mysql!(query, params)
 }
